# docs/API/openapi-schema.yaml
openapi: 3.0.3
info:
  title: ApexStore Mini E-Commerce & Order Management API
  version: 1.0.0
  description: Comprehensive REST API documentation for ApexStore e-commerce platform.
    Covers Authentication, Products, Categories, Cart, Wishlist, Coupons, Orders,
    Payments, Inventory, and Admin Analytics.
paths:
...
  /api/orders/{id}/:
    get:
      operationId: orders_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - orders
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
          description: ''
...
  /api/orders/{id}/status/:
    patch:
      operationId: orders_status_partial_update
      description: 'Admin: Update order status (and handle inventory return if cancelled)'
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - orders
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
components:
  schemas:
...
    Coupon:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        code:
          type: string
          maxLength: 50
        description:
          type: string
        discount_type:
          $ref: '#/components/schemas/DiscountTypeEnum'
        discount_value:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        min_order_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        max_discount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          nullable: true
        start_date:
          type: string
          format: date-time
        expiration_date:
          type: string
          format: date-time
        usage_limit:
          type: integer
          maximum: 4294967295
          minimum: 0
          format: int64
          nullable: true
          description: Total system usage limit
        per_user_usage_limit:
          type: integer
          maximum: 4294967295
          minimum: 0
          format: int64
        is_active:
          type: boolean
        total_used:
          type: integer
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - code
      - created_at
      - discount_value
      - expiration_date
      - id
      - total_used
      - updated_at
...
    DiscountTypeEnum:
      enum:
      - PERCENTAGE
      - FIXED_AMOUNT
      type: string
      description: |-
        * `PERCENTAGE` - Percentage
        * `FIXED_AMOUNT` - Fixed Amount
...
    Order:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        order_number:
          type: string
          readOnly: true
        customer:
          type: integer
          readOnly: true
        customer_username:
          type: string
          readOnly: true
        customer_email:
          type: string
          readOnly: true
        subtotal:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
        discount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
        shipping_fee:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
        total_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
        coupon:
          type: integer
          nullable: true
        coupon_details:
          allOf:
          - $ref: '#/components/schemas/Coupon'
          readOnly: true
        order_status:
          $ref: '#/components/schemas/OrderStatusEnum'
        payment_status:
          $ref: '#/components/schemas/PaymentStatusEnum'
        shipping_full_name:
          type: string
          maxLength: 150
        shipping_phone:
          type: string
          maxLength: 20
        shipping_address:
          type: string
        shipping_city:
          type: string
          maxLength: 100
        shipping_district:
          type: string
          maxLength: 100
        shipping_postal_code:
          type: string
          maxLength: 20
        shipping_country:
          type: string
          maxLength: 100
        notes:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
          readOnly: true
        payment:
          allOf:
          - $ref: '#/components/schemas/Payment'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - coupon_details
      - created_at
      - customer
      - customer_email
      - customer_username
      - discount
      - id
      - items
      - order_number
      - payment
      - shipping_address
      - shipping_city
      - shipping_district
      - shipping_fee
      - shipping_full_name
      - shipping_phone
      - shipping_postal_code
      - subtotal
      - total_amount
      - updated_at
    OrderItem:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        product:
          type: integer
        product_name:
          type: string
          maxLength: 255
        product_sku:
          type: string
          maxLength: 50
        unit_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        quantity:
          type: integer
          maximum: 4294967295
          minimum: 0
          format: int64
        subtotal:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
      required:
      - id
      - product
      - product_name
      - product_sku
      - quantity
      - subtotal
      - unit_price
    OrderStatusEnum:
      enum:
      - PENDING
      - CONFIRMED
      - PROCESSING
      - SHIPPED
      - DELIVERED
      - CANCELLED
      type: string
      description: |-
        * `PENDING` - Pending
        * `CONFIRMED` - Confirmed
        * `PROCESSING` - Processing
        * `SHIPPED` - Shipped
        * `DELIVERED` - Delivered
        * `CANCELLED` - Cancelled
...
    Payment:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        order:
          type: integer
        transaction_id:
          type: string
          maxLength: 100
        payment_method:
          $ref: '#/components/schemas/PaymentMethodEnum'
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        status:
          $ref: '#/components/schemas/PaymentStatusEnum'
        payment_details: {}
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - amount
      - created_at
      - id
      - order
      - transaction_id
    PaymentMethodEnum:
      enum:
      - CARD
      - MOBILE_BANKING
      - CASH_ON_DELIVERY
      type: string
      description: |-
        * `CARD` - Credit / Debit Card (Mock)
        * `MOBILE_BANKING` - Mobile Banking (Mock)
        * `CASH_ON_DELIVERY` - Cash on Delivery (Mock)
    PaymentStatusEnum:
      enum:
      - PENDING
      - PAID
      - FAILED
      - REFUNDED
      type: string
      description: |-
        * `PENDING` - Pending
        * `PAID` - Paid
        * `FAILED` - Failed
        * `REFUNDED` - Refunded
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT